BigDummy/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1